+++ src/CVNBot/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SWMTBot")]
+[assembly: AssemblyTitle("CVNBot")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SWMTBot")]
-[assembly: AssemblyCopyright("Copyright © 2006-2007 Tangotango. All rights reserved.")]
+[assembly: AssemblyProduct("CVNBot")]
+[assembly: AssemblyCopyright("Copyright (c) 2006-2012 Countervandalism Network and other contributors.")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]